+++ spec/codecs/cef_spec.rb
@@ -304,6 +304,49 @@ describe LogStash::Codecs::CEF do
       insist { e.get('severity') } == "10"
     end
 
+    ##
+    # Use the given codec to decode the given data, ensuring exactly one event is emitted.
+    #
+    # If a block is given, yield the resulting event to the block _outside_ of `LogStash::Codecs::CEF#decode(String)`
+    # in order to avoid mismatched-exceptions raised by RSpec triggering the codec's exception-handling.
+    #
+    # @param codec [#decode]
+    # @param data [String]
+    # @yieldparam event [Event]
+    # @yieldreturn [void]
+    # @return [Event]
+    def decode_one(codec, data)
+      events = do_decode(codec, data)
+      fail("Expected one event, got #{events.size} events: #{events.inspect}") unless events.size == 1
+      event = events.first
+
+      yield event if block_given?
+
+      event
+    end
+
+    ##
+    # Use the given codec to decode the given data, returning an Array of the resulting Events
+    #
+    # If a block is given, each event is yielded to the block _outside_ of `LogStash::Codecs::CEF#decode(String)`
+    # in order to avoid mismatched-exceptions raised by RSpec triggering the codec's exception-handling.
+    #
+    # @param codec [#decode]
+    # @param data [String]
+    # @yieldparam event [Event]
+    # @yieldreturn [void]
+    # @return [Array<Event>]
+    def do_decode(codec, data)
+      events = []
+      codec.decode(data) do |event|
+        events << event
+      end
+
+      events.each { |event| yield event } if block_given?
+
+      events
+    end
+
     context "with delimiter set" do
       # '\r\n' in single quotes to simulate the real input from a config
       # containing \r\n as 4-character sequence in the config:
@@ -314,24 +357,36 @@ describe LogStash::Codecs::CEF do
       subject(:codec) { LogStash::Codecs::CEF.new("delimiter" => '\r\n') }
 
       it "should parse on the delimiter " do
-        subject.decode(message) do |e|
+        do_decode(subject,message) do |e|
           raise Exception.new("Should not get here. If we do, it means the decoder emitted an event before the delimiter was seen?")
         end
 
-        event = false;
-        subject.decode("\r\n") do |e|
+        decode_one(subject, "\r\n") do |e|
           validate(e)
           insist { e.get("deviceVendor") } == "security"
           insist { e.get("deviceProduct") } == "threatmanager"
-          event = true
         end
+      end
+    end
 
-        expect(event).to be_truthy
+    context 'when a CEF header ends with a pair of properly-escaped backslashes' do
+      let(:backslash) { '\\' }
+      let(:pipe) { '|' }
+      let(:message) { "CEF:0|security|threatmanager|1.0|100|double backslash" +
+                      backslash + backslash + # escaped backslash
+                      backslash + backslash + # escaped backslash
+                      "|10|src=10.0.0.192 dst=12.121.122.82 spt=1232" }
+
+      it 'should include the backslashes unescaped' do
+        event = decode_one(subject, message)
+
+        expect(event.get('name')).to eq('double backslash' + backslash + backslash )
+        expect(event.get('severity')).to eq('10') # ensure we didn't consume the separator
       end
     end
 
     it "should parse the cef headers" do
-      subject.decode(message) do |e|
+      decode_one(subject, message) do |e|
         validate(e)
         insist { e.get("deviceVendor") } == "security"
         insist { e.get("deviceProduct") } == "threatmanager"
@@ -339,7 +394,7 @@ describe LogStash::Codecs::CEF do
     end
 
     it "should parse the cef body" do
-      subject.decode(message) do |e|
+      decode_one(subject, message) do |e|
         insist { e.get("sourceAddress")} == "10.0.0.192"
         insist { e.get("destinationAddress") } == "12.121.122.82"
         insist { e.get("sourcePort") } == "1232"
@@ -348,7 +403,7 @@ describe LogStash::Codecs::CEF do
 
     let (:missing_headers) { "CEF:0|||1.0|100|trojan successfully stopped|10|src=10.0.0.192 dst=12.121.122.82 spt=1232" }
     it "should be OK with missing CEF headers (multiple pipes in sequence)" do
-      subject.decode(missing_headers) do |e|
+      decode_one(subject, missing_headers) do |e|
         validate(e)
         insist { e.get("deviceVendor") } == ""
         insist { e.get("deviceProduct") } == ""
@@ -357,35 +412,79 @@ describe LogStash::Codecs::CEF do
 
     let (:leading_whitespace) { "CEF:0|security|threatmanager|1.0|100|trojan successfully stopped|10| src=10.0.0.192 dst=12.121.122.82 spt=1232" }
     it "should strip leading whitespace from the message" do
-      subject.decode(leading_whitespace) do |e|
+      decode_one(subject, leading_whitespace) do |e|
         validate(e)
       end
     end
 
     let (:escaped_pipes) { 'CEF:0|security|threatmanager|1.0|100|trojan successfully stopped|10|moo=this\|has an escaped pipe' }
     it "should be OK with escaped pipes in the message" do
-      subject.decode(escaped_pipes) do |e|
+      decode_one(subject, escaped_pipes) do |e|
         insist { e.get("moo") } == 'this\|has an escaped pipe'
       end
     end
 
     let (:pipes_in_message) {'CEF:0|security|threatmanager|1.0|100|trojan successfully stopped|10|moo=this|has an pipe'}
     it "should be OK with not escaped pipes in the message" do
-      subject.decode(pipes_in_message) do |e|
+      decode_one(subject, pipes_in_message) do |e|
         insist { e.get("moo") } == 'this|has an pipe'
       end
     end
 
+    # while we may see these in practice, equals MUST be escaped in the extensions per the spec.
     let (:equal_in_message) {'CEF:0|security|threatmanager|1.0|100|trojan successfully stopped|10|moo=this =has = equals\='}
     it "should be OK with equal in the message" do
-      subject.decode(equal_in_message) do |e|
+      decode_one(subject, equal_in_message) do |e|
         insist { e.get("moo") } == 'this =has = equals='
+      end
+    end
+
+    let(:malformed_unescaped_equals_in_extension_value) { %q{CEF:0|FooBar|Web Gateway|1.2.3.45.67|200|Success|2|rt=Sep 07 2018 14:50:39 cat=Access Log dst=1.1.1.1 dhost=foo.example.com suser=redacted src=2.2.2.2 requestMethod=POST request='https://foo.example.com/bar/bingo/1' requestClientApplication='Foo-Bar/2018.1.7; Email:user@example.com; Guid:test=' cs1= cs1Label=Foo Bar} }
+    it 'should split correctly' do
+      decode_one(subject, malformed_unescaped_equals_in_extension_value) do |event|
+        expect(event.get('cefVersion')).to eq('0')
+        expect(event.get('deviceVendor')).to eq('FooBar')
+        expect(event.get('deviceProduct')).to eq('Web Gateway')
+        expect(event.get('deviceVersion')).to eq('1.2.3.45.67')
+        expect(event.get('deviceEventClassId')).to eq('200')
+        expect(event.get('name')).to eq('Success')
+        expect(event.get('severity')).to eq('2')
+
+        # extension key/value pairs
+        expect(event.get('deviceReceiptTime')).to eq('Sep 07 2018 14:50:39')
+        expect(event.get('deviceEventCategory')).to eq('Access Log')
+        expect(event.get('deviceVersion')).to eq('1.2.3.45.67')
+        expect(event.get('destinationAddress')).to eq('1.1.1.1')
+        expect(event.get('destinationHostName')).to eq('foo.example.com')
+        expect(event.get('sourceUserName')).to eq('redacted')
+        expect(event.get('sourceAddress')).to eq('2.2.2.2')
+        expect(event.get('requestMethod')).to eq('POST')
+        expect(event.get('requestUrl')).to eq(%q{'https://foo.example.com/bar/bingo/1'})
+        # Although the value for `requestClientApplication` contains an illegal unquoted equals sign, the sequence
+        # preceeding the unescaped-equals isn't shaped like a key, so we allow it to be a part of the value.
+        expect(event.get('requestClientApplication')).to eq(%q{'Foo-Bar/2018.1.7; Email:user@example.com; Guid:test='})
+        expect(event.get('deviceCustomString1Label')).to eq('Foo Bar')
+        expect(event.get('deviceCustomString1')).to eq('')
+      end
+    end
+
+    context('escaped-equals and unescaped-spaces in the extension values') do
+      let(:query_string) { 'key1=value1&key2=value3 aa.bc&key3=value4'}
+      let(:escaped_query_string) { query_string.gsub('=','\\=') }
+      let(:cef_message) { "CEF:0|security|threatmanager|1.0|100|trojan successfully stopped|10|go=start now query_string=#{escaped_query_string} final=done" }
+
+      it 'captures the extension values correctly' do
+        event = decode_one(subject, cef_message)
+
+        expect(event.get('go')).to eq('start now')
+        expect(event.get('query_string')).to eq(query_string)
+        expect(event.get('final')).to eq('done')
       end
     end
 
     let (:escaped_backslash_in_header) {'CEF:0|secu\\\\rity|threat\\\\manager|1.\\\\0|10\\\\0|tro\\\\jan successfully stopped|\\\\10|'}
     it "should be OK with escaped backslash in the headers" do
-      subject.decode(escaped_backslash_in_header) do |e|
+      decode_one(subject, escaped_backslash_in_header) do |e|
         insist { e.get("cefVersion") } == '0'
         insist { e.get("deviceVendor") } == 'secu\\rity'
         insist { e.get("deviceProduct") } == 'threat\\manager'
@@ -398,7 +497,7 @@ describe LogStash::Codecs::CEF do
 
     let (:escaped_backslash_in_header_edge_case) {'CEF:0|security\\\\\\||threatmanager\\\\|1.0|100|trojan successfully stopped|10|'}
     it "should be OK with escaped backslash in the headers (edge case: escaped slash in front of pipe)" do
-      subject.decode(escaped_backslash_in_header_edge_case) do |e|
+      decode_one(subject, escaped_backslash_in_header_edge_case) do |e|
         validate(e)
         insist { e.get("deviceVendor") } == 'security\\|'
         insist { e.get("deviceProduct") } == 'threatmanager\\'
@@ -407,7 +506,7 @@ describe LogStash::Codecs::CEF do
 
     let (:escaped_pipes_in_header) {'CEF:0|secu\\|rity|threatmanager\\||1.\\|0|10\\|0|tro\\|jan successfully stopped|\\|10|'}
     it "should be OK with escaped pipes in the headers" do
-      subject.decode(escaped_pipes_in_header) do |e|
+      decode_one(subject, escaped_pipes_in_header) do |e|
         insist { e.get("cefVersion") } == '0'
         insist { e.get("deviceVendor") } == 'secu|rity'
         insist { e.get("deviceProduct") } == 'threatmanager|'
@@ -420,14 +519,14 @@ describe LogStash::Codecs::CEF do
 
     let (:backslash_in_message) {'CEF:0|security|threatmanager|1.0|100|trojan successfully stopped|10|moo=this \\has \\ backslashs\\'}
     it "should be OK with backslashs in the message" do
-      subject.decode(backslash_in_message) do |e|
+      decode_one(subject, backslash_in_message) do |e|
         insist { e.get("moo") } == 'this \\has \\ backslashs\\'
       end
     end
 
     let (:equal_in_header) {'CEF:0|security|threatmanager=equal|1.0|100|trojan successfully stopped|10|'}
     it "should be OK with equal in the headers" do
-      subject.decode(equal_in_header) do |e|
+      decode_one(subject, equal_in_header) do |e|
         validate(e)
         insist { e.get("deviceProduct") } == "threatmanager=equal"
       end
@@ -435,7 +534,7 @@ describe LogStash::Codecs::CEF do
 
     let (:spaces_in_between_keys) {'CEF:0|security|threatmanager|1.0|100|trojan successfully stopped|10| src=10.0.0.192  dst=12.121.122.82  spt=1232'}
     it "should be OK to have one or more spaces between keys" do
-      subject.decode(spaces_in_between_keys) do |e|
+      decode_one(subject, spaces_in_between_keys) do |e|
         validate(e)
         insist { e.get("sourceAddress") } == "10.0.0.192"
         insist { e.get("destinationAddress") } == "12.121.122.82"
@@ -445,7 +544,7 @@ describe LogStash::Codecs::CEF do
 
     let (:allow_spaces_in_values) {'CEF:0|security|threatmanager|1.0|100|trojan successfully stopped|10|src=10.0.0.192 dst=12.121.122.82  spt=1232 dproc=InternetExplorer x.x.x.x'}
     it "should be OK to have one or more spaces in values" do
-      subject.decode(allow_spaces_in_values) do |e|
+      decode_one(subject, allow_spaces_in_values) do |e|
         validate(e)
         insist { e.get("sourceAddress") } == "10.0.0.192"
         insist { e.get("destinationAddress") } == "12.121.122.82"
@@ -456,26 +555,26 @@ describe LogStash::Codecs::CEF do
 
     let (:preserve_additional_fields_with_dot_notations) {'CEF:0|security|threatmanager|1.0|100|trojan successfully stopped|10|src=10.0.0.192 additional.dotfieldName=new_value ad.Authentification=MICROSOFT_AUTHENTICATION_PACKAGE_V1_0 ad.Error_,Code=3221225578 dst=12.121.122.82 ad.field[0]=field0 ad.name[1]=new_name'}
     it "should keep ad.fields" do
-      subject.decode(preserve_additional_fields_with_dot_notations) do |e|
+      decode_one(subject, preserve_additional_fields_with_dot_notations) do |e|
         validate(e)
         insist { e.get("sourceAddress") } == "10.0.0.192"
         insist { e.get("destinationAddress") } == "12.121.122.82"
-        insist { e.get("ad.field[0]") } == "field0"
-        insist { e.get("ad.name[1]") } == "new_name"
+        insist { e.get("[ad.field][0]") } == "field0"
+        insist { e.get("[ad.name][1]") } == "new_name"
         insist { e.get("ad.Authentification") } == "MICROSOFT_AUTHENTICATION_PACKAGE_V1_0"
-        insist { e.get("ad.Error_,Code") } == "3221225578"
+        insist { e.get('ad.Error_,Code') } == "3221225578"
         insist { e.get("additional.dotfieldName") } == "new_value"
       end
     end
 
     let (:preserve_random_values_key_value_pairs_alongside_with_additional_fields) {'CEF:0|security|threatmanager|1.0|100|trojan successfully stopped|10|src=10.0.0.192 cs4=401 random.user Admin 0 23041A10181C0000  23041810181C0000  /CN\=random.user/OU\=User Login End-Entity  /CN\=TEST/OU\=Login CA TEST 34 additional.dotfieldName=new_value ad.Authentification=MICROSOFT_AUTHENTICATION_PACKAGE_V1_0 ad.Error_,Code=3221225578 dst=12.121.122.82 ad.field[0]=field0 ad.name[1]=new_name'}
     it "should correctly parse random values even with additional fields in message" do
-      subject.decode(preserve_random_values_key_value_pairs_alongside_with_additional_fields) do |e|
+      decode_one(subject, preserve_random_values_key_value_pairs_alongside_with_additional_fields) do |e|
         validate(e)
         insist { e.get("sourceAddress") } == "10.0.0.192"
         insist { e.get("destinationAddress") } == "12.121.122.82"
-        insist { e.get("ad.field[0]") } == "field0"
-        insist { e.get("ad.name[1]") } == "new_name"
+        insist { e.get("[ad.field][0]") } == "field0"
+        insist { e.get("[ad.name][1]") } == "new_name"
         insist { e.get("ad.Authentification") } == "MICROSOFT_AUTHENTICATION_PACKAGE_V1_0"
         insist { e.get("ad.Error_,Code") } == "3221225578"
         insist { e.get("additional.dotfieldName") } == "new_value"
@@ -485,7 +584,7 @@ describe LogStash::Codecs::CEF do
 
     let (:preserve_unmatched_key_mappings) {'CEF:0|security|threatmanager|1.0|100|trojan successfully stopped|10|src=10.0.0.192 dst=12.121.122.82 new_key_by_device=new_values here'}
     it "should preserve unmatched key mappings" do
-      subject.decode(preserve_unmatched_key_mappings) do |e|
+      decode_one(subject, preserve_unmatched_key_mappings) do |e|
         validate(e)
         insist { e.get("sourceAddress") } == "10.0.0.192"
         insist { e.get("destinationAddress") } == "12.121.122.82"
@@ -495,7 +594,7 @@ describe LogStash::Codecs::CEF do
 
     let (:translate_abbreviated_cef_fields) {'CEF:0|security|threatmanager|1.0|100|trojan successfully stopped|10|src=10.0.0.192 dst=12.121.122.82 proto=TCP shost=source.host.name dhost=destination.host.name spt=11024 dpt=9200 outcome=Success amac=00:80:48:1c:24:91'}
     it "should translate most known abbreviated CEF field names" do
-      subject.decode(translate_abbreviated_cef_fields) do |e|
+      decode_one(subject, translate_abbreviated_cef_fields) do |e|
         validate(e)
         insist { e.get("sourceAddress") } == "10.0.0.192"
         insist { e.get("destinationAddress") } == "12.121.122.82"
@@ -511,9 +610,42 @@ describe LogStash::Codecs::CEF do
 
     let (:syslog) { "Syslogdate Sysloghost CEF:0|security|threatmanager|1.0|100|trojan successfully stopped|10|src=10.0.0.192 dst=12.121.122.82 spt=1232" }
     it "Should detect headers before CEF starts" do
-      subject.decode(syslog) do |e|
+      decode_one(subject, syslog) do |e|
         validate(e)
         insist { e.get('syslog') } == 'Syslogdate Sysloghost'
+      end
+    end
+
+    context 'with UTF-8 message' do
+      let(:message) { 'CEF:0|security|threatmanager|1.0|100|trojan successfully stopped|10|src=192.168.1.11 target=aaaaaああああaaaa msg=Description Omitted' }
+
+      # since this spec is encoded UTF-8, the literal strings it contains are encoded with UTF-8,
+      # but codecs in Logstash tend to receive their input as BINARY (or: ASCII-8BIT); ensure that
+      # we can handle either without losing the UTF-8 characters from the higher planes.
+      %w(
+        BINARY
+        UTF-8
+      ).each do |external_encoding|
+        context "externally encoded as #{external_encoding}" do
+          let(:message) { super().force_encoding(external_encoding) }
+          it 'should keep the higher-plane characters' do
+            decode_one(subject, message.dup) do |event|
+              validate(event)
+              insist { event.get("target") } == "aaaaaああああaaaa"
+              insist { event.get("target").encoding } == Encoding::UTF_8
+            end
+          end
+        end
+      end
+    end
+
+    context 'non-UTF-8 message' do
+      let(:message) { 'CEF:0|security|threatmanager|1.0|100|trojan successfully stopped|10|src=192.168.1.11 target=aaaaaああああaaaa msg=Description Omitted'.encode('SHIFT_JIS') }
+      it 'should emit message unparsed with _cefparsefailure tag' do
+        decode_one(subject, message.dup) do |event|
+          insist { event.get("message").bytes.to_a } == message.bytes.to_a
+          insist { event.get("tags") } == ['_cefparsefailure']
+        end
       end
     end
 
@@ -521,7 +653,7 @@ describe LogStash::Codecs::CEF do
       subject(:codec) { LogStash::Codecs::CEF.new("raw_data_field" => "message_raw") }
 
       it "should return the raw message in field message_raw" do
-        subject.decode(message) do |e|
+        decode_one(subject, message.dup) do |e|
           validate(e)
           insist { e.get("message_raw") } == message
         end